task/serializers.py
from rest_framework import serializers
from .models import Task
from django.contrib.auth.models import User


class TaskSerializer(serializers.ModelSerializer):  # create class to serializer model
    creator = serializers.ReadOnlyField(source='creator.username')
    class Meta:
        model = Task
        fields = ('id', 'taskdetail', 'creator')


class UserSerializer(serializers.ModelSerializer):  # create class to serializer user model
    tasks = serializers.PrimaryKeyRelatedField(many=True, queryset=Task.objects.all())

    class Meta:
        model = User
        fields = ('id', 'username', 'tasks')
