requirements.txt.txt
asgiref
Django==3.2.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytz==2021.1
sqlparse==0.4.1